added configurations for remaining subsystems and organized the folders
