+++ Team498/src/main/java/frc/robot/configurations/ClampSubsystemConfiguration.java
@@ -1,0 +1,17 @@
+/*----------------------------------------------------------------------------*/
+/* Copyright (c) 2018 FIRST. All Rights Reserved.                             */
+/* Open Source Software - may be modified and shared by FRC teams. The code   */
+/* must be accompanied by the FIRST BSD license file in the root directory of */
+/* the project.                                                               */
+/*----------------------------------------------------------------------------*/
+
+package frc.robot.configurations;
+
+/**
+ * Add your docs here.
+ */
+public class ClampSubsystemConfiguration {
+    //TODO: change these channels
+    public static final int clampForwardChannel = 0;
+    public static final int clampReverseChannel = 0;
+}
