+++ Team498/src/main/java/frc/robot/configurations/ControllerConfiguration.java
@@ -5,10 +5,10 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot;
+package frc.robot.configurations;
 
 
-public class Mappings {
+public class ControllerConfiguration {
 	//controller port(s)
 	public static final int ControllerPort = 0;
 	
@@ -23,6 +23,7 @@ public class Mappings {
 	public static final int ButtonStart = 8;
 	public static final int LeftJoyPress = 9;
 	public static final int RightJoyPress = 10;
+	
 
 	//triggers and joysticks 
 	public static final int LeftXAxis = 0;
@@ -32,14 +33,5 @@ public class Mappings {
 	public static final int RightTrigger = 3;
 	public static final int LeftTrigger = 2;
 
-	//motor channels
-	//TODO: need to change these motor channels
-	public static final int frontLeftDriveMotorChannel = 0;
-	public static final int backLeftDriveMotorChannel = 0; 
-	public static final int frontRightDriveMotorChannel = 0;
-	public static final int backRightDriveMotorChannel = 0;
-
-
-	//TODO: need to change this motor channel
-	public static final int wristMotorChannel = 0;
+	
 }
