+++ Team498/src/main/java/frc/robot/configurations/DrivetimeConfiguration.java
@@ -1,0 +1,20 @@
+/*----------------------------------------------------------------------------*/
+/* Copyright (c) 2018 FIRST. All Rights Reserved.                             */
+/* Open Source Software - may be modified and shared by FRC teams. The code   */
+/* must be accompanied by the FIRST BSD license file in the root directory of */
+/* the project.                                                               */
+/*----------------------------------------------------------------------------*/
+
+package frc.robot.configurations;
+
+/**
+ * Add your docs here.
+ */
+public class DrivetimeConfiguration {
+	//TODO: need to change these motor channels
+	public static final int frontLeftDriveMotorChannel = 0;
+	public static final int backLeftDriveMotorChannel = 0; 
+	public static final int frontRightDriveMotorChannel = 0;
+    public static final int backRightDriveMotorChannel = 0;
+    
+}
