+++ Team498/src/main/java/frc/robot/configurations/WristSubsystemConfiguration.java
@@ -1,0 +1,14 @@
+/*----------------------------------------------------------------------------*/
+/* Copyright (c) 2018 FIRST. All Rights Reserved.                             */
+/* Open Source Software - may be modified and shared by FRC teams. The code   */
+/* must be accompanied by the FIRST BSD license file in the root directory of */
+/* the project.                                                               */
+/*----------------------------------------------------------------------------*/
+
+package frc.robot.configurations;
+
+
+public class WristSubsystemConfiguration {
+	//TODO: need to change this motor channel
+	public static final int wristMotorChannel = 0;
+}
