+++ Team498/src/main/java/frc/robot/subsystems/Drivetrain.java
@@ -12,7 +12,7 @@ import edu.wpi.first.wpilibj.SpeedControllerGroup;
 import edu.wpi.first.wpilibj.command.Subsystem;
 import edu.wpi.first.wpilibj.drive.DifferentialDrive;
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
-import frc.robot.subsystems.DrivetimeConfiguration;
+import frc.robot.configurations.DrivetimeConfiguration;
 import frc.robot.commands.ManualDrive;
