+++ Team498/src/main/java/frc/robot/subsystems/WristSubsystem.java
@@ -9,7 +9,7 @@ package frc.robot.subsystems;
 
 
 import edu.wpi.first.wpilibj.command.Subsystem;
-import frc.robot.Mappings;
+import frc.robot.configurations.WristSubsystemConfiguration;
 import frc.robot.commands.ManualWristCommand;
 import com.ctre.phoenix.motorcontrol.can.WPI_VictorSPX;
 
@@ -23,7 +23,7 @@ public class WristSubsystem extends Subsystem {
   private static WristSubsystem wristSubsystem = null;
 
   //TODO: Need to change motor control channel :3
-  private WPI_VictorSPX wrist = new WPI_VictorSPX(Mappings.wristMotorChannel);
+  private WPI_VictorSPX wrist = new WPI_VictorSPX(WristSubsystemConfiguration.wristMotorChannel);
